requirements.txt
Django
djangorestframework
psycopg2